requirements.txt
appdirs==1.4.4
apple-music-python==1.0.3
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.8.2
bleach==3.2.1
browser-cookie3==0.12.0
bs4==0.0.1
certifi==2020.11.8
cffi==1.14.0
chardet==4.0.0
click==7.1.2
cPython==0.0.6
cryptography==2.8
cv==1.0.0
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
docutils==0.16
emoji==0.5.4
entrypoints==0.3
ExifRead==2.3.1
filelock==3.0.12
Flask==1.1.2
future==0.18.2
gpsphoto==2.2.3
greenlet==0.4.17
h5py==3.1.0
idna==2.10
imageio==2.9.0
imutils==0.5.4
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.17.2
jeepney==0.6.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keyring==21.8.0
kiwisolver==1.3.1
lz4==3.1.2
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
msgpack==1.0.2
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
neovim==0.3.1
nest-asyncio==1.4.3
networkx==2.5
notebook==6.2.0
numpy==1.19.4
opencv-contrib-python==4.4.0.46
opencv-python==4.4.0.40
packaging==20.8
pandas==1.1.5
pandocfilters==1.4.3
-e git+https://github.com/guettli/panorama.git@4fef2c2b7d081bbf745644989ae319ebeed33e1b#egg=panorama
parso==0.7.1
pbkdf2==1.3
pexpect==4.8.0
pickleshare==0.7.5
piexif==1.1.3
Pillow==8.0.1
prometheus-client==0.9.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pyaes==1.6.1
pycparser==2.20
pycryptodome==3.9.9
Pygments==2.7.3
PyJWT==1.7.1
pymongo==3.11.0
pynvim==0.4.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
Python-dev==2.0.0.dev0
python-resize-image==1.1.19
python-telegram-bot==12.5
pytz==2020.4
PyWavelets==1.1.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
requests==2.25.1
scipy==1.5.4
SecretStorage==3.3.0
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0
telegram==0.0.1
terminado==0.9.2
testpath==0.4.4
tifffile==2020.11.18
tornado==6.1
traitlets==5.0.5
urllib3==1.26.3
virtualenv==20.2.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
